+++ NEWS.md
@@ -7,6 +7,7 @@
 * Added roxygen documentation for all package data objects
 * Added `@param` entries for `clusters` and `filter` in the `ft` class
 * Added testthat coverage for all alpha-diversity functions
+* Added embed argument to `obs2gfm()` function
 
 # jakR2 0.3